Various fixes and clean up
- set_unicode_preference bug fix
- minor typo fixes
- removed some unused imports
- fixed the api setting sync
- moved unicode code to new file
