+++ bot/setup/update_settings_from_api.py
@@ -4,10 +4,10 @@ from fractions import Fraction
 
 async def update_settings_from_api():
     try:
-        settings = await api_client.get_settings
+        settings = api_client.get_settings()
         config.MAXIMUM_DEBT_CHARACTER_LIMIT = int(settings.get("MAXIMUM_DEBT_CHARACTER_LIMIT", config.MAXIMUM_DEBT_CHARACTER_LIMIT))
         config.MAXIMUM_PER_DEBT = int(settings.get("MAXIMUM_PER_DEBT", config.MAXIMUM_PER_DEBT))
-        config.SMALLEST_UNIT = Fraction(("SMALLEST_UNIT", config.SMALLEST_UNIT))
+        config.SMALLEST_UNIT = Fraction(settings.get("SMALLEST_UNIT", config.SMALLEST_UNIT))
         config.QUANTIZE_SETTLING_DEBTS = bool(settings.get("QUANTIZE_SETTLING_DEBTS", config.QUANTIZE_SETTLING_DEBTS))
         config.QUANTIZE_OWING_DEBTS = bool(settings.get("QUANTIZE_OWING_DEBTS", config.QUANTIZE_OWING_DEBTS))
         config.SORT_OWES_FIRST = bool(settings.get("SORT_OWES_FIRST", config.SORT_OWES_FIRST))
